More testing

+++ Sockets/WebServer/src/main/java/funHttpServer/WebServer.java
@@ -253,7 +253,7 @@ class WebServer {
               JSONObject jsonObject = (JSONObject) o;
               JSONObject owner = (JSONObject) jsonObject.get("owner");
               String loginName = (String) owner.get("login");
-              Integer ownerId = (Integer) owner.get("id");
+              Long ownerId = (Long) owner.get("id");
               String repoName = (String) jsonObject.get("name");
               System.out.print(loginName + ", " + ownerId + "-> " + repoName);
               //System.out.println(repoName);
